# cli/api/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"HttpExceptionBody": {
				"properties": {
					"message": {
						"$ref": "#/components/schemas/HttpExceptionBodyMessage"
					},
					"error": {
						"type": "string"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"message",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"RawTransaction": {
				"$ref": "#/components/schemas/SdkRawTransaction"
			},
...
			"PoolInfo": {
				"description": "General Pool Interface",
				"properties": {
					"aValue": {
						"type": "string",
						"description": "Pool A value"
					},
					"dValue": {
						"type": "string",
						"description": "Pool D value"
					},
					"tokenBalance": {
						"type": "string",
						"description": "Pool token balance"
					},
					"vUsdBalance": {
						"type": "string",
						"description": "Pool virtual USD balance"
					},
					"totalLpAmount": {
						"type": "string",
						"description": "Pool total liquidity amount"
					},
					"accRewardPerShareP": {
						"type": "string",
						"description": "Current accumulated reward per share shifted by P bits"
					},
					"p": {
						"type": "number",
						"format": "double",
						"description": "P value, needed for accRewardPerShareP"
					},
					"imbalance": {
						"type": "string",
						"description": "Pool imbalance"
					}
				},
				"required": [
					"aValue",
					"dValue",
					"tokenBalance",
					"vUsdBalance",
					"totalLpAmount",
					"accRewardPerShareP",
					"p",
					"imbalance"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "rest-api",
		"version": "3.30.0-alpha.2",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/raw/approve": {
			"get": {
				"operationId": "Approve",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RawTransaction"
								},
								"examples": {
									"EVM approve transaction": {
										"value": {
											"from": "0x0000000000000000000000000000000000000000",
											"to": "0x0000000000000000000000000000000000000000",
											"value": "1000",
											"data": "0x0000000000000000000000000000000000000000000000000000000000000000"
										}
									},
									"Tron approve transaction": {
										"value": {
											"visible": false,
											"txID": "0000000000000000000000000000000000000000000000000000000000000000",
											"raw_data": {
												"contract": [],
												"ref_block_bytes": "0000",
												"ref_block_hash": "0000000000000000",
												"expiration": 0,
												"timestamp": 0
											},
											"raw_data_hex": "00"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Pool",
					"Transfers",
					"Raw Transactions",
					"Yield"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "ownerAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tokenAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "amount",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"default": "bridge",
							"type": "string",
							"enum": [
								"bridge",
								"pool",
								"yield"
							]
						}
					},
					{
						"in": "query",
						"name": "messenger",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"ALLBRIDGE",
								"WORMHOLE",
								"CCTP",
								"CCTP_V2",
								"OFT",
								"X_RESERVE"
							]
						}
					},
					{
						"in": "query",
						"name": "feePaymentMethod",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"WITH_NATIVE_CURRENCY",
								"WITH_STABLECOIN",
								"WITH_ABR"
							]
						}
					},
					{
						"in": "query",
						"name": "contractAddress",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/raw/pool/approve": {
			"get": {
				"operationId": "PoolApprove",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RawTransaction"
								},
								"examples": {
									"Pool approve raw transaction": {
										"value": "0xffff"
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Pool",
					"Raw Transactions"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "ownerAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tokenAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "amount",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/raw/deposit": {
			"get": {
				"operationId": "Deposit",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RawTransaction"
								},
								"examples": {
									"Pool deposit raw transaction": {
										"value": "0xffff"
									},
									"Stacks pool deposit transaction": {
										"value": "80800000000400b168de3a000000000000000100000000000000000000000000000000"
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Pool",
					"Raw Transactions"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "amount",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "ownerAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tokenAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "solanaTxFeeParams",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"AUTO",
								"PRICE_PER_UNIT_IN_MICRO_LAMPORTS",
								"EXTRA_FEE_IN_LAMPORTS"
							]
						}
					},
					{
						"in": "query",
						"name": "solanaTxFeeValue",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/raw/withdraw": {
			"get": {
				"operationId": "Withdraw",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RawTransaction"
								},
								"examples": {
									"Pool withdraw raw transaction": {
										"value": "0xffff"
									},
									"Stacks pool withdraw transaction": {
										"value": "80800000000400b168de3a000000000000000100000000000000000000000000000000"
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Pool",
					"Raw Transactions"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "amount",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "ownerAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tokenAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "solanaTxFeeParams",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"AUTO",
								"PRICE_PER_UNIT_IN_MICRO_LAMPORTS",
								"EXTRA_FEE_IN_LAMPORTS"
							]
						}
					},
					{
						"in": "query",
						"name": "solanaTxFeeValue",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/raw/claim": {
			"get": {
				"operationId": "ClaimRewards",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RawTransaction"
								},
								"examples": {
									"Pool claim rewards raw transaction": {
										"value": "0xffff"
									},
									"Stacks pool claim rewards transaction": {
										"value": "80800000000400b168de3a000000000000000100000000000000000000000000000000"
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Pool",
					"Raw Transactions"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "ownerAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tokenAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "solanaTxFeeParams",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"AUTO",
								"PRICE_PER_UNIT_IN_MICRO_LAMPORTS",
								"EXTRA_FEE_IN_LAMPORTS"
							]
						}
					},
					{
						"in": "query",
						"name": "solanaTxFeeValue",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/check/pool/allowance": {
			"get": {
				"operationId": "CheckPoolAllowance",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								},
								"examples": {
									"Pool allowance check result": {
										"value": true
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Tokens",
					"Pool"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "amount",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "ownerAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tokenAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "feePaymentMethod",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"WITH_NATIVE_CURRENCY",
								"WITH_STABLECOIN",
								"WITH_ABR"
							]
						}
					}
				]
			}
		},
		"/pool/info/server": {
			"get": {
				"operationId": "GetPoolInfoByServer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PoolInfo"
								},
								"examples": {
									"Pool info from server": {
										"value": {
											"aValue": "100",
											"dValue": "2000000",
											"tokenBalance": "1500000",
											"vUsdBalance": "1499000",
											"totalLpAmount": "3000000",
											"accRewardPerShareP": "42",
											"p": 48,
											"imbalance": "0.0001"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Pool"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "tokenAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/pool/info/blockchain": {
			"get": {
				"operationId": "GetPoolInfoFromBlockchain",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Required_PoolInfo_"
								},
								"examples": {
									"Pool info from blockchain": {
										"value": {
											"aValue": "100",
											"dValue": "2000000",
											"tokenBalance": "1500000",
											"vUsdBalance": "1499000",
											"totalLpAmount": "3000000",
											"accRewardPerShareP": "42",
											"p": 48,
											"imbalance": "0.0001"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Pool"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "tokenAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/pool/allowance": {
			"get": {
				"operationId": "GetPoolAllowance",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								},
								"examples": {
									"Pool allowance amount": {
										"value": "100000000"
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Pool",
					"Tokens"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "ownerAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tokenAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "feePaymentMethod",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"WITH_NATIVE_CURRENCY",
								"WITH_STABLECOIN",
								"WITH_ABR"
							]
						}
					}
				]
			}
		},
		"/liquidity/details": {
			"get": {
				"operationId": "GetUserPoolInfo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserLiquidityDetails"
								},
								"examples": {
									"User liquidity position": {
										"value": {
											"lpAmount": "12.5",
											"rewardDebt": "0.03"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Pool"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "ownerAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tokenAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/liquidity/deposit/calculate": {
			"get": {
				"operationId": "GetAmountToBeDeposited",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								},
								"examples": {
									"Estimated LP amount on deposit": {
										"value": "12.345678"
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Pool"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "amount",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tokenAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/liquidity/withdrawn/calculate": {
			"get": {
				"operationId": "GetAmountToBeWithdrawn",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								},
								"examples": {
									"Estimated token amount on withdraw": {
										"value": "10.125"
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Pool"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "amount",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "ownerAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tokenAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/liquidity/vusd/calculate": {
			"get": {
				"operationId": "GetVUsdFromAmount",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AmountFormatted"
								},
								"examples": {
									"vUsd amount for the provided token amount": {
										"value": {
											"int": "1000000",
											"float": "1"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Pool"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "amount",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "amountFormat",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"INT",
								"FLOAT"
							]
						}
					},
					{
						"in": "query",
						"name": "tokenAddress",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}